Keep AST traversal lightweight for GD&T defaults
