+++ src/lang/queryAst.ts
@@ -1,7 +1,6 @@
 import type { FunctionExpression } from '@rust/kcl-lib/bindings/FunctionExpression'
 import type { ImportStatement } from '@rust/kcl-lib/bindings/ImportStatement'
 import type { Node } from '@rust/kcl-lib/bindings/Node'
-import type { TypeDeclaration } from '@rust/kcl-lib/bindings/TypeDeclaration'
 import {
   createLiteral,
   createLocalName,
@@ -32,14 +31,11 @@ import type {
   BinaryExpression,
   CallExpressionKw,
   Expr,
-  ExpressionStatement,
-  Identifier,
   Literal,
   Name,
   PathToNode,
   PipeExpression,
   Program,
-  ReturnStatement,
   SegmentArtifact,
   SourceRange,
   SyntaxType,
@@ -61,7 +57,7 @@ import type { Artifact, Plane } from '@rust/kcl-lib/bindings/Artifact'
 import type { NumericType } from '@rust/kcl-lib/bindings/NumericType'
 import type { OpArg, Operation } from '@rust/kcl-lib/bindings/Operation'
 import type { SketchBlock } from '@rust/kcl-lib/bindings/SketchBlock'
-import { ARG_INDEX_FIELD, LABELED_ARG_FIELD } from '@src/lang/queryAstConstants'
+import { traverse } from '@src/lang/traverse'
 import type { KclCommandValue } from '@src/lib/commandTypes'
 import type { ModuleType } from '@src/lib/wasm_lib_wrapper'
 import type {
@@ -70,6 +66,8 @@ import type {
   Selections,
 } from '@src/machines/modelingSharedTypes'
 import type { UnaryExpression } from 'typescript'
+
+export { traverse } from '@src/lang/traverse'
 
 /**
  * Retrieves a node from a given path within a Program node structure, optionally stopping at a specified node type.
@@ -221,118 +219,6 @@ export function getNodeFromPathCurry(
       path: shallowPath,
     }
   }
-}
-
-type KCLNode = Node<
-  | Expr
-  | ExpressionStatement
-  | ImportStatement
-  | VariableDeclaration
-  | VariableDeclarator
-  | TypeDeclaration
-  | ReturnStatement
-  | Identifier
->
-
-export function traverse(
-  node: KCLNode | Node<Program>,
-  option: {
-    enter?: (node: KCLNode, pathToNode: PathToNode) => void
-    leave?: (node: KCLNode) => void
-  },
-  pathToNode: PathToNode = []
-) {
-  const _node = node as KCLNode
-  option?.enter?.(_node, pathToNode)
-  const _traverse = (node: KCLNode, pathToNode: PathToNode) =>
-    traverse(node, option, pathToNode)
-
-  if (_node.type === 'VariableDeclaration') {
-    _traverse(_node.declaration, [
-      ...pathToNode,
-      ['declaration', 'VariableDeclaration'],
-    ])
-  } else if (_node.type === 'VariableDeclarator') {
-    _traverse(_node.init, [...pathToNode, ['init', '']])
-  } else if (_node.type === 'ExpressionStatement') {
-    _traverse(_node.expression, [
-      ...pathToNode,
-      ['expression', 'ExpressionStatement'],
-    ])
-  } else if (_node.type === 'PipeExpression') {
-    _node.body.forEach((expression, index) =>
-      _traverse(expression, [
-        ...pathToNode,
-        ['body', 'PipeExpression'],
-        [index, 'index'],
-      ])
-    )
-  } else if (_node.type === 'CallExpressionKw') {
-    _traverse(_node.callee, [...pathToNode, ['callee', 'CallExpressionKw']])
-    if (_node.unlabeled !== null) {
-      _traverse(_node.unlabeled, [
-        ...pathToNode,
-        ['unlabeled', 'Unlabeled arg'],
-      ])
-    }
-    if (_node.arguments) {
-      _node.arguments.forEach((arg, index) =>
-        _traverse(arg.arg, [
-          ...pathToNode,
-          ['arguments', 'CallExpressionKw'],
-          [index, ARG_INDEX_FIELD],
-          ['arg', LABELED_ARG_FIELD],
-        ])
-      )
-    }
-  } else if (_node.type === 'BinaryExpression') {
-    _traverse(_node.left, [...pathToNode, ['left', 'BinaryExpression']])
-    _traverse(_node.right, [...pathToNode, ['right', 'BinaryExpression']])
-  } else if (_node.type === 'Name') {
-    // do nothing
-  } else if (_node.type === 'Literal') {
-    // do nothing
-  } else if (_node.type === 'TagDeclarator') {
-    // do nothing
-  } else if (_node.type === 'ArrayExpression') {
-    _node.elements.forEach((el, index) =>
-      _traverse(el, [
-        ...pathToNode,
-        ['elements', 'ArrayExpression'],
-        [index, 'index'],
-      ])
-    )
-  } else if (_node.type === 'ObjectExpression') {
-    _node.properties.forEach(({ key, value }, index) => {
-      _traverse(key, [
-        ...pathToNode,
-        ['properties', 'ObjectExpression'],
-        [index, 'index'],
-        ['key', 'Property'],
-      ])
-      _traverse(value, [
-        ...pathToNode,
-        ['properties', 'ObjectExpression'],
-        [index, 'index'],
-        ['value', 'Property'],
-      ])
-    })
-  } else if (_node.type === 'UnaryExpression') {
-    _traverse(_node.argument, [...pathToNode, ['argument', 'UnaryExpression']])
-  } else if (_node.type === 'MemberExpression') {
-    // hmm this smell
-    _traverse(_node.object, [...pathToNode, ['object', 'MemberExpression']])
-    _traverse(_node.property, [...pathToNode, ['property', 'MemberExpression']])
-  } else if (_node.type === 'ImportStatement') {
-    // Do nothing.
-  } else if ('body' in _node && isArray(_node.body)) {
-    // TODO: Program should have a type field, but it currently doesn't.
-    const program = node as Node<Program>
-    program.body.forEach((expression, index) => {
-      _traverse(expression, [...pathToNode, ['body', ''], [index, 'index']])
-    })
-  }
-  option?.leave?.(_node)
 }
 
 export interface PrevVariable<T> {
